requirements.txt
opencv-python
streamlit
transformers
torch
pillow
